refactor(classifier): folder names and content
