--- petools/model_tools/pose_classifier/process_data/utils/__init__.py
@@ -1,3 +1,0 @@
-
-from .feature_generator import FeatureGenerator
-from .constants import conlist
